Fix MESSAGE_CONTAINS

+++ src/main/java/com/game/service/impl/ImageService.java
@@ -37,7 +37,7 @@ public class ImageService implements IImageService
     private ImageRepository imageRepository;
     @Autowired
     private GameRepository gameRepository;
-    private static ImageConverter imageConverter = ImageConverter.getInstance();
+    private final ImageConverter imageConverter = ImageConverter.getInstance();
     @Override
     @Transactional
     public ImageDto save(MultipartFile file , Integer gameId , Integer mapValue)
@@ -75,17 +75,17 @@ public class ImageService implements IImageService
                                        String orderBy, String sortDir, Integer page, Integer limit)
     {
         if (page == null || limit == null)
-            throw APIException.from(HttpStatus.BAD_REQUEST).withMessage(MessageConstants.Page_And_Limit_Not_Null);
+            throw APIException.from(HttpStatus.BAD_REQUEST).withMessage(MessageConstants.PAGE_AND_LIMIT_NOT_NULL);
         Pageable pageable = PageRequest.of(page-1,limit);
         Set<User> users = gameRepository.findUsersByGameId(gameId);
         if (!SecurityUtils.getInstance().isGamePlayer(users))
-            throw APIException.from(HttpStatus.FORBIDDEN).withMessage(MessageConstants.Not_Gamer);
+            throw APIException.from(HttpStatus.FORBIDDEN).withMessage(MessageConstants.NOT_GAMER);
         if (orderBy != null && sortDir != null)
             pageable = PageRequest.of(pageable.getPageNumber(), pageable.getPageNumber(),
                     Sort.by(Sort.Direction.valueOf(sortDir),orderBy));
         List<Image> images = imageRepository.findAll(gameId,active,activePlay,pageable);
         if (images == null || images.isEmpty())
-            throw APIException.from(HttpStatus.NOT_FOUND).withMessage(MessageConstants.Image_Not_Found);
+            throw APIException.from(HttpStatus.NOT_FOUND).withMessage(MessageConstants.IMAGE_NOT_FOUND);
         return imageConverter.toDto(images);
     }
 
@@ -94,15 +94,14 @@ public class ImageService implements IImageService
     public List<ImageDto> findAllByActivePlay(Integer gameId) {
         Set<User> users = gameRepository.findUsersByGameId(gameId);
             if (!SecurityUtils.getInstance().isGamePlayer(users))
-                throw APIException.from(HttpStatus.FORBIDDEN).withMessage(MessageConstants.Not_Gamer);
+                throw APIException.from(HttpStatus.FORBIDDEN).withMessage(MessageConstants.NOT_GAMER);
         List<Object> images = imageRepository.getLinkImages(gameId);
         if (images == null || images.isEmpty())
-            throw  APIException.from(HttpStatus.NOT_FOUND).withMessage(MessageConstants.Not_Found);
+            throw  APIException.from(HttpStatus.NOT_FOUND).withMessage(MessageConstants.NOT_FOUND);
         List<ImageDto> result = new ArrayList<>();
-
-        for (int i=0;i<images.size();i++)
-        {
-            Object[] object = (Object[]) images.get(i);
+        // just show id , link , mapValue for user
+        for (Object image : images) {
+            Object[] object = (Object[]) image;
             ImageDto imageDto = ImageDto.from((String) object[2])
                     .value((Integer) object[1])
                     .id((Integer) object[0]);
@@ -131,11 +130,14 @@ public class ImageService implements IImageService
     }
 
     @Override
+    @Transactional
     public void delete(Integer[] ids) {
         for (Integer id : ids) {
             String link = imageRepository.getLinkImage(id);
             File file = new File ("src/main/resources/static/"+link);
             boolean remove = file.delete();
+            if (!remove)
+                throw APIException.from(HttpStatus.NOT_ACCEPTABLE).withMessage(MessageConstants.IMAGE_REMOVE_FAILED);
             imageRepository.deleteById(id);
         }
     }
